project/fastapi/requirements.txt
fastapi==0.103.1
python-dotenv==1.0.0
SQLAlchemy==2.0.20
uvicorn==0.23.2
async-sqlalchemy==1.0.0
dependency-injector==4.41.0